# PERF: Optimize StringExtensions.cs